Create SAS code
this is the SAS code that sets up the macro and applies it to an example from NSSRN data. Caveat is that the computation is complex and takes time, with 128 GB RAM and Intel(R) Core(TM) i9-9900 CPU @ 3.10GHz, i believe it took around 3 hours for 80 replicates (+1 for full sample) and 1000 burn-in MCMC, 10000 MCMC iterations with thinning at 10 for 1000 posterior draws associated with each replicate. I'm going to look into parallization to speed up run and may further augment by having it loop over a set of dependent variables and perhaps send it to a more high performance scientific computing system.